refactor: GlobalProvider분리
