+++ src/components/RestaurantList.tsx
@@ -1,6 +1,6 @@
-import React, { useContext } from "react";
-import { RestaurantContext } from "../App";
-import RestaurantItem from "./RestaurantItem";
+import React, { useContext } from 'react';
+import { RestaurantContext } from '../containers/GlobalProvider';
+import RestaurantItem from './RestaurantItem';
 
 const RestaurantList = () => {
   const { state } = useContext(RestaurantContext);
@@ -10,7 +10,7 @@ const RestaurantList = () => {
     <ul>
       {restaurants
         .filter((restaurant) =>
-          categorizeBy === "all" ? true : restaurant.category === categorizeBy
+          categorizeBy === 'all' ? true : restaurant.category === categorizeBy
         )
         .map((restaurant) => (
           <RestaurantItem key={restaurant.name} restaurant={restaurant} />
